backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2022.7.1
sqlparse==0.4.3
Pillow==9.5.0
requests==2.30.0
drf-yasg==1.21.5
channels==3.0.4